pyproject.toml
[project]
name = "eeg-mi-bci"
version = "0.1.0"
description = "EEG-based Motor Imagery Brain-Computer Interface system"
requires-python = ">=3.8"
dependencies = [
    "mne>=1.0.0",
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "pylsl>=1.16.0",
    "joblib>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "moabb>=1.0.0",
]

[project.optional-dependencies]
deep = ["torch>=1.9.0"]
dev = ["pytest>=7.0.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
